Bring back parse as built-in.
