+++ crates/nu-cli/src/commands/parse/mod.rs
@@ -1,0 +1,3 @@
+mod command;
+
+pub use command::Command as Parse;
