--- crates/nu_plugin_parse/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "nu_plugin_parse"
-version = "0.14.1"
-authors = ["The Nu Project Contributors"]
-edition = "2018"
-description = "A string parsing plugin for Nushell"
-license = "MIT"
-
-[lib]
-doctest = false
-
-[dependencies]
-nu-plugin = { path = "../nu-plugin", version = "0.14.1" }
-nu-protocol = { path = "../nu-protocol", version = "0.14.1" }
-nu-source = { path = "../nu-source", version = "0.14.1" }
-nu-errors = { path = "../nu-errors", version = "0.14.1" }
-futures = { version = "0.3", features = ["compat", "io-compat"] }
-regex = "1"
-
-[build-dependencies]
-nu-build = { version = "0.14.1", path = "../nu-build" }
