--- crates/nu_plugin_parse/build.rs
@@ -1,3 +1,0 @@
-fn main() -> Result<(), Box<dyn std::error::Error>> {
-    nu_build::build()
-}
