--- crates/nu_plugin_parse/src/lib.rs
@@ -1,4 +1,0 @@
-mod nu;
-mod parse;
-
-pub use parse::Parse;
