--- crates/nu_plugin_parse/src/parse.rs
@@ -1,21 +1,0 @@
-use nu_source::Tag;
-use regex::Regex;
-
-pub struct Parse {
-    pub regex: Regex,
-    pub name: Tag,
-    pub pattern_tag: Tag,
-    pub column_names: Vec<String>,
-}
-
-impl Parse {
-    #[allow(clippy::trivial_regex)]
-    pub fn new() -> Result<Self, Box<dyn std::error::Error>> {
-        Ok(Parse {
-            regex: Regex::new("")?,
-            name: Tag::unknown(),
-            pattern_tag: Tag::unknown(),
-            column_names: vec![],
-        })
-    }
-}
